requirements.txt
fastmcp>=2.3.4
sqlalchemy>=2.0.41
pyodbc>=5.2.0
pydantic>=2.11.5
pydantic-settings>=2.9.1
redis>=5.3.0
sqlglot>=26.16.4
python-dotenv>=1.1.0
tenacity>=9.1.2
structlog>=25.3.0
fastapi>=0.115.0
uvicorn>=0.34.0